The android studio project folders
Github wouldnt let me upload them as one zip because it's 100+ mb so i had winrar break it into volumes of 24mb each. I think you just need to download the 4 .z0# files and the zip, then when you open the zip it will see the portions if they are in the same directory. The github should now hold all the files necessary to continue this endeavor should anyone choose to do so!